Refactor worker reply building into a request_reply module
